move 'main' method to separate class
