+++ src/main/java/island/Island.java
@@ -42,27 +42,27 @@ public class Island {
         return totalAnimalsCount > 0;
     }
 
-    public static void main(String[] args) {
-//        int islandWidth = Integer.parseInt(args[0]);
-//        int islandHeight = Integer.parseInt(args[1]);
-
-        int islandWidth = 3;
-        int islandHeight = 3;
-
-        Island theIsland = new Island(islandWidth, islandHeight);
-        int roundNumber = 1;
-
-        System.out.println("\n\n******************************************************************* LIFE BEGUN **********************************************************");
-        while (theIsland.isEnaughAnimals()) {
-            System.out.println("******************************************************************* ROUND " + roundNumber + " *************************************************************");
-            for (int i = 0; i < theIsland.islandWidth; i++) {
-                for (int j = 0; j < theIsland.islandHeight; j++) {
-                    cells[i][j].lifeRound();
-                    cells[i][j].reportStatus();
-                }
-            }
-            roundNumber++;
-        }
-
-    }
+//    public static void main(String[] args) {
+////        int islandWidth = Integer.parseInt(args[0]);
+////        int islandHeight = Integer.parseInt(args[1]);
+//
+//        int islandWidth = 3;
+//        int islandHeight = 3;
+//
+//        Island theIsland = new Island(islandWidth, islandHeight);
+//        int roundNumber = 1;
+//
+//        System.out.println("\n\n******************************************************************* LIFE BEGUN **********************************************************");
+//        while (theIsland.isEnaughAnimals()) {
+//            System.out.println("******************************************************************* ROUND " + roundNumber + " *************************************************************");
+//            for (int i = 0; i < theIsland.islandWidth; i++) {
+//                for (int j = 0; j < theIsland.islandHeight; j++) {
+//                    cells[i][j].lifeRound();
+//                    cells[i][j].reportStatus();
+//                }
+//            }
+//            roundNumber++;
+//        }
+//
+//    }
 }
